+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
-### v0.3.0 - 2024-05-xx
+
+### v0.4.0 - 2024-06-xx
+
+- `lexicmap search`:
+    - Perform more accurate alignment with [WFA](https://github.com/shenwei356/wfa).
+
+### v0.3.0 - 2024-05-14
 
 - `lexicmap index`:
     - **Better seed coverage by filling sketching deserts**.